# src/main/java/org/example/neekostar/bank/dto/UserUpdateDto.java
package org.example.neekostar.bank.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class UserUpdateDto {

    @NotBlank(message = "Firstname is required")
    private String firstName;

    @NotBlank(message = "Lastname is required")
    private String lastName;
}